refactor: create enum for openid and extract
