+++ packages/common/src/index.ts
@@ -47,6 +47,10 @@ import { SlackSettings } from './types/slackSettings';
 import { EmailStatusExpiring } from './types/email';
 import { FieldValueSearchResult } from './types/fieldMatch';
 import {
+    DeleteOpenIdentity,
+    OpenIdIdentitySummary,
+} from './types/openIdIdentity';
+import {
     AllowedEmailDomains,
     OnboardingStatus,
     Organization,
@@ -98,6 +102,7 @@ export * from './types/fieldMatch';
 export * from './types/filter';
 export * from './types/job';
 export * from './types/metricQuery';
+export * from './types/openIdIdentity';
 export * from './types/organization';
 export * from './types/organizationMemberProfile';
 export * from './types/personalAccessToken';
@@ -382,33 +387,6 @@ export type UpdateUserArgs = {
     isTrackingAnonymized: boolean;
     isSetupComplete: boolean;
 };
-
-export type CreateOpenIdIdentity = {
-    subject: string;
-    issuer: string;
-    issuerType: 'google' | 'okta' | 'oneLogin';
-    userId: number;
-    email: string;
-};
-
-export type UpdateOpenIdentity = Pick<
-    CreateOpenIdIdentity,
-    'subject' | 'issuer' | 'email' | 'issuerType'
->;
-
-export type OpenIdIdentity = CreateOpenIdIdentity & {
-    createdAt: Date;
-};
-
-export type OpenIdIdentitySummary = Pick<
-    OpenIdIdentity,
-    'issuer' | 'email' | 'createdAt' | 'issuerType'
->;
-
-export type DeleteOpenIdentity = Pick<
-    OpenIdIdentitySummary,
-    'issuer' | 'email'
->;
 
 export type PasswordResetLink = {
     expiresAt: Date;
